refactor the code to be more robust
